+++ tests/Integration/FactoryTest.php
@@ -11,6 +11,7 @@ use DjThossi\ErgosoftSdk\Api\GetJobByGuidApi;
 use DjThossi\ErgosoftSdk\Api\GetJobsApi;
 use DjThossi\ErgosoftSdk\Api\SubscribeJobStatusApi;
 use DjThossi\ErgosoftSdk\Api\SubmitDeltaXmlFileApi;
+use DjThossi\ErgosoftSdk\Api\UnsubscribeJobStatusApi;
 use DjThossi\ErgosoftSdk\Domain\BaseUrl;
 use DjThossi\ErgosoftSdk\ErgosoftFactory;
 use DjThossi\ErgosoftSdk\SimpleErgosoftConfiguration;
@@ -72,6 +73,14 @@ class FactoryTest extends TestCase
 
         /* @noinspection UnnecessaryAssertionInspection */
         $this->assertInstanceOf(SubscribeJobStatusApi::class, $factory->createSubscribeJobStatusApi());
+    }
+
+    public function testCreateUnsubscribeJobStatusApi(): void
+    {
+        $factory = $this->createErgosoftFactory();
+
+        /* @noinspection UnnecessaryAssertionInspection */
+        $this->assertInstanceOf(UnsubscribeJobStatusApi::class, $factory->createUnsubscribeJobStatusApi());
     }
 
     private function createErgosoftFactory(): ErgosoftFactory